feat(miner-foundation): extract reward-risk scoring into gittensory-engine (#2281)
Move the reward/risk reasoning signals into packages/gittensory-engine/src/reward-risk.ts
so the gittensory-miner can rank candidate work locally with the same logic the
maintainer-side gate computes.

Closes #2281.
